feat: move config to top level script
